pyproject.toml
[tool.poetry]
name = "django+aiogram"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"
pre-commit = "^4.0.1"
mypy = "^1.13.0"


[tool.poetry.group.back.dependencies]
django = "^5.1.3"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
drf-spectacular = "^0.27.2"
drf-yasg = {extras = ["validation"], version = "^1.21.8"}


[tool.poetry.group.bot.dependencies]
aiogram = "^3.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "single"